==> Notebooks/00.eda.ipynb <==
# About Dataset

### **Context**
This Online Retail II dataset contains all the transactions occurring for a UK-based and registered, non-store online retail between 01.12.2009 and 09.12.2011. The company mainly sells unique all-occasion gift-ware. Many customers of the company are wholesalers.

The dataset is available on Kaggle (https://www.kaggle.com/datasets/kabilan45/online-retail-ii-datase) and on UCI Machine Learning Repository (https://archive.ics.uci.edu/dataset/502/online+retail+ii).

### **Content**
#### **Attribute Information:**

- `Invoice`: Invoice number. Nominal. A 6-digit integral number uniquely assigned to each transaction. If this code starts with the letter 'c', it indicates a cancellation.
- `StockCode`: Product (item) code. Nominal. A 5-digit integral number uniquely assigned to each distinct product.
- `Description`: Product (item) name. Nominal.
- `Quantity`: The quantities of each product (item) per transaction. Numeric.
- `InvoiceDate`: Invoice date and time. Numeric. The day and time when a transaction was generated.
- `Price`: Unit price. Numeric. Product price per unit in sterling (Â£).
- `Customer ID`: Customer number. Nominal. A 5-digit integral number uniquely assigned to each customer.
- `Country`: Country name. Nominal. The name of the country where a customer resides.
---

# 1. Getting to know our Dataset

# Import libraries
import numpy as np 
import pandas as pd 
import datetime as dt
import matplotlib.pyplot as plt 
from IPython.display import display
from collections import defaultdict

# Read excel's two sheets and merge them into a single Dataframe
df1_raw = pd.read_excel("online_retail_II.xlsx", sheet_name = "Year 2009-2010")
df2_raw = pd.read_excel("online_retail_II.xlsx", sheet_name = "Year 2010-2011")
df_raw = pd.concat([df1_raw, df2_raw])

# Review info on columns
df_raw.info()

# Review first rows of dataset
df_raw.head()

# Shape of merged dataset
df_raw.shape

It looks like we have **8 columns** with **1067371 rows**. 

### **1.1 Check for duplicates**


# Check for duplicates
print("Number of duplicate rows: ", df_raw.sort_values(by = "InvoiceDate").duplicated().sum())

# Visually inspect duplicate data records
_duplicated = df_raw[df_raw.duplicated(keep = False)].sort_values(by = ["InvoiceDate", "Invoice", "StockCode"])
_duplicated

It is fair to suspect that some of the **34335** duplicate records occurred due to merging data from the two excel sheets (*Years 2009-2010*, *Years 2010-2011*). We review the final date of the ***df1_raw*** and the first date of the ***df2_raw*** in order to confirm this assumption.

# Review the final date of the *df1_raw* and the first date of the *df2_raw*.   

df1_raw_final_date = df1_raw["InvoiceDate"].max()
df2_raw_start_date = df2_raw["InvoiceDate"].min()

print("Final date of the 'Year 2009-2010' dataset:", df1_raw_final_date)
print("Start date of the 'Year 2010-2011' dataset:", df2_raw_start_date)

It appears there is a data overlap of 9 days of transactions between the two excel sheets.
This can also be seen visually by filtering **df1_raw** by InvoiceDate, i.e. *InvoiceDate>=01.12.2010* and printing the first few rows of the resulted dataframe, while also printing the first few rows of **df2_raw**.

# First few rows from 12.2010 from the sheet 'Year 2009-2010'
df1_raw[df1_raw["InvoiceDate"] >= dt.datetime(2010,12,1)].sort_values(by = "InvoiceDate").head()

# First rows from the sheet 'Year 2010-2011'
df2_raw.head()

We remove the overlap in the dataset and create a new dataframe named **df**.

# Remove overlapped records, create df 
df = pd.concat([df1_raw[df1_raw["InvoiceDate"] < df2_raw_start_date], df2_raw])

df.shape

 print("Number of duplicate rows:", df.sort_values(by = "InvoiceDate").duplicated().sum())

duplicated = df[df.duplicated(keep = False)].sort_values(by = ["InvoiceDate", "Invoice", "StockCode"])
duplicated

# Export df to excel in order to load in PowerBI
# df.to_excel(r'C:\Users\argyr\OneDrive\Υπολογιστής\Diplvm\Giftware dataset\online_retail_II_removed duplicates.xlsx', index=False)

# List columns and see data types
df.info()

# Take a closer look at column distributions
df.describe()

### **1.2 Identify Missing Values**


# Identify missing values
missing = df.isnull().sum()
print(missing)

# Percentage of missing values in each column
missing/len(df) * 100

---
# 2. Exploratory Data Analysis

### 2.1 Invoice Column

# Convert column Invoice's data type to string
df["Invoice"] = df["Invoice"].astype(str)

# Unique number of invoices
len(df["Invoice"].unique())

# Create a DataFrame with the records of cancelled invoices
df_cancelled_inv = df[df["Invoice"].str.contains("C")]
df_cancelled_inv.head()

# Review dimensions of cancelled invoices DataFrame
df_cancelled_inv.shape

The transactions that contain cancelled invoices (Invoice starts with "C") are **19165**. Out of these 19165 transactions, only 1 has positive value on Quantity column, while the other 19164 Invoices that start with "C" also have negative values on Quantity column.

# Invoices that start with "C" and Quantity < 0
df_cancelled_inv[df_cancelled_inv["Quantity"] < 0]

# Invoices that start with "C" and Quantity > 0
df_cancelled_inv[df_cancelled_inv["Quantity"] > 0]

Moreover, there are some records where Invoice starts with "**A**". Let's try to figure out the meaning of these records.

# Invoices that start with "A"
df[df["Invoice"].str.contains("A")]

df[df["Invoice"].str.contains("A")].shape

There are only 6 transactions of what appears to be adjustments regarding debts of the company. This explains the strongly negative values on **Price** column. There is only one exception regarding Invoice "A563185" with *Price = 11062.06* (positive value). Invoice *A563185* seems to be related with *A563186* and *A563187*, although they cannot be characterised as duplicates.

# Invoice starts with "C" and Customer ID is blank
df_cancelled_inv["Customer ID"].isnull().sum()

### 2.2 Quantity Column

# Number of records where Quantity < 0 
len(df[(df["Quantity"] < 0)])

There are **22557** records with negative value on Quantity. These records are more than the cancelled invoices with negative value on Quantity (19164), which means that negative quantity can be linked to other kinds of transactions too, not only cancelled invoices/returns of products.

df["Quantity"].isnull().sum()

# Number of records where Quantity < 0 and Customer ID is blank
df[(df["Quantity"] < 0) & df["Customer ID"].isnull()]

# Number of records where Quantity < 0, Customer ID is blank and Invoice starts with "C"
df_cancelled_inv[(df_cancelled_inv["Quantity"] < 0) & df_cancelled_inv["Customer ID"].isnull()]

### 2.3 Customer ID Column

# Number of records where Customer ID is blank 
len(df[(df["Customer ID"].isnull())])

Customer ID column refers to the unique ID number given to each customer. There are **235287** missing values in this column, namely **22.5%** of the total number of transactions (1044848). This proportion is of great importance, and given the chance we would ask of the data collectors to explain this gap in the dataset. Given the fact that this is not possible, we assume that these customers had not registered to the online shop and thus, were never assigned a customer ID. 

# Number of unique customers 
unique_customers = len(df["Customer ID"].unique())
print("There are {} unique Customer IDs in the dataset.".format(unique_customers - 1))

We substract 1, which represents the distinct count of null values in Customer ID column. 

# Drop missing values in Customer ID (235 287)
df_nnz_customerID = df.dropna(subset = ["Customer ID"])
print("For the customer segmentation analysis that will follow, we will drop missing values from " 
      "Customer ID column and thus the shape of the DataFrame will be reduced to {}.".format(df_nnz_customerID.shape))

### 2.4 Country Column

# List of countries 
countries_list = df["Country"].unique()
number_of_countries = len(df["Country"].unique())
print("The number of countries in the dataset is {} and they are listed below:\n\n".format(number_of_countries), countries_list)

# Number of records where country is "Unspecified"
len(df[df["Country"] == "Unspecified"])

# Top 10 countries by number of transactions
print(df["Country"].value_counts()[:10]/len(df) * 100)

A large proportion of the customers buy from United Kingdom (**91.8%**), which is also the base of the online shop.

### 2.5 StockCode Column

# Number of unique StockCode items
number_of_StockCodes = pd.unique(df["StockCode"])
print("The number of unique StockCodes is", len(number_of_StockCodes))

According to Python, the unique values of the column StockCode is **5305**.
In parallel with Python we conducted the same analysis using Microsoft Power BI, where we constructed a measure using the DAX expression *DISTINCTCOUNT* in order to count the unique values of the column. The result in Power BI is **5131**. We tried to figure out which of these results was the right one.
It's reasonable to assume that Python internally regards certain values as distinct whereas Power BI does not. Otherwise, this discrepancy wouldn't arise. In order to investigate whether this suspicion is correct, we must first level the playing field with regard to the values of the column. To this end, we cast all values as type string and subsequently turn all strings to upppercase.  In this way all values are comparable with each other.
After the execution of these commands, we see that the unique values of column StockCode are **5132**!

len(df['StockCode'].astype(str).str.upper().sort_values().unique())

In order to find that one value that remains as a difference, we exported both lists of unique StockCodes to excel. The value *47503J* was spotted as written with two different ways, namely **"47503J "** and **"47503J"**. We used the command *str.strip()* in order to substract the blank space that caused this and finally reached the conclusion that the unique values of column StockCode are **5131**.

# df_2222 = pd.DataFrame(df['StockCode'].astype(str).str.upper().sort_values().unique())
# df_2222.to_excel(r'C:\Users\argyr\OneDrive\Υπολογιστής\Diplvm\Giftware dataset\online_retail_II_stockcodepy.xlsx', index=False)

(df['StockCode'].astype(str).str.upper().sort_values().unique() == '47503J').sum()

(df['StockCode'].astype(str).str.upper().sort_values().unique() == '47503J ').sum()

Using Power BI, we didn't have to perform all these steps to reach the right number of unique StockCodes because the command *DISTINCTCOUNTNOBLANK* counts the unique values of a column ignoring the fact that there may be blank spaces or uppercase/lowercase letters in value names. While here this has been helpful,it is not always the case. Each dataset is unique and it is absolutely necessary for the analyst to deeply understand its peculiarities before heading on with (and trusting?) any analysis tool.

We create a new DataFrame (**df2**) in order to store the values of *StockCode* in their updated version. From now on, **df2** is the base for our analysis.

# New version of df. This is the base version for the next steps of the analysis.
df['StockCode'] = df['StockCode'].astype(str).str.upper().str.strip()
df2 = df.copy()

unique_stockcodes = df['StockCode'].unique()
len(unique_stockcodes)

#### Top 10 StockCodes

df2["StockCode"].isnull().sum()

# We would like to subtract records whose Price is zero.
# However Python and Power BI do not seem to agree on the representation of zero.
# Therefore in order to determine which products should be excluded from the top ten analysis,
# we will find out which products have positive price, which have negative, and then subtract
# these from the lot.

df_price_positive = df2[df2["Price"] > 0.0]
df_price_positive.shape

df_price_negative = df2[df2["Price"] < 0.0]
df_price_negative.shape

df_price_zero = df2[df2["Price"] == 0]
df_price_zero.shape

# The above three sizes agree with those extracted from Power BI.

# We concatenate the two DataFrames.
df_nnz_prices = pd.concat([df_price_positive, df_price_negative])

# We group by StockCode and sum the values of Quantity in order to review the ten most popular StockCodes of the online shop.
df_fixed_stockcodes = df_nnz_prices.copy()
most_preferred_products = df_fixed_stockcodes.groupby(["StockCode"])["Quantity"].sum().sort_values(ascending=False)[:10]
print(most_preferred_products)
most_preferred_products = pd.DataFrame(most_preferred_products)

top10_stockcodes = most_preferred_products.index.tolist()
print(top10_stockcodes)

# Print the descriptions of the ten most popular StockCodes
for i in range(len(top10_stockcodes)):
    print(df[df["StockCode"] == top10_stockcodes[i]]["Description"][:1])

### 2.6 InvoiceDate Column

The column **InvoiceDate** refers to the date and time that a transaction was generated. The transactions range from 01.12.2009 to 09.12.2011, as described at the start of the analysis.

# There are only 5 records whose price is negative. These transactions, as we have mentioned above,
# are purely logistical and do not reflect sales. 

# Our goal now is to find out which month has the most sales. 
# We restrict our analysis only to products whose price is positive.
df3 = df_price_positive.copy()

earliest_invoice_date = df3["InvoiceDate"].min()
latest_invoice_date = df3["InvoiceDate"].max()
date_fmt = "%d-%m-%Y"
print("The data ranges from {} to {}.".format(earliest_invoice_date.strftime(date_fmt), 
                                             latest_invoice_date.strftime(date_fmt) 
                                            )
     )

# Convert InvoiceDate column to datetime format
df3["InvoiceDate"] = pd.to_datetime(df3["InvoiceDate"])

# Create Month column
df3["Month"] = df3["InvoiceDate"].dt.strftime("%B")

# Create TotalAmount column
df3["TotalAmount"] = df3["Price"] * df3["Quantity"] 

# Monthly sales
monthly_sales = df3.groupby(["Month"])["TotalAmount"].sum().sort_values(ascending=False)
print(monthly_sales)

# Our goal now is to find out which hour of the day has the most distinct orders. 
# We restrict our analysis only to products whose price is positive.

# Create Hour column
df3["Hour"] = df3["InvoiceDate"].dt.strftime("%H")

# Number of orders 
hourly_invoices = df3.groupby(["Hour"])["Invoice"].nunique()
print(hourly_invoices)
