==> Notebooks/01.cleandata.ipynb <==
## Data Cleaning

# Import libraries
import numpy as np 
import pandas as pd 
import datetime as dt
import matplotlib.pyplot as plt 
from IPython.display import display
from collections import defaultdict

# Merge the two spreadsheets
df1_raw = pd.read_excel("online_retail_II.xlsx", sheet_name = "Year 2009-2010")
df1_raw.reset_index()
df2_raw = pd.read_excel("online_retail_II.xlsx", sheet_name = "Year 2010-2011")
df_raw = pd.concat([df1_raw, df2_raw])
df_raw.shape

# Remove 9 days overlap
df2_raw_start_date = df2_raw["InvoiceDate"].min()
df = pd.concat([df1_raw[df1_raw["InvoiceDate"] < df2_raw_start_date], df2_raw])
df.shape 

# Column datatype conversions: Invoice to string 
df["Invoice"] = df["Invoice"].astype(str)
df.shape

# Column datatype conversions: StockCode to string, uppercase letters, substract blank space from value names
df["StockCode"] = df["StockCode"].astype(str).str.upper().str.strip()
df.shape

# Drop missing values in Customer ID (235287)
df = df.dropna(subset = ["Customer ID"])
df.shape

# Shape of df before dropping transactions with negative quantities
rows_before_negq_drop = df.shape[0]
df.shape

# Number of transactions with negative quantities
len(df[df["Quantity"] < 0])

# Create new column 'CIDxSC Summed Quantity' in which for each combination of Customer ID and StockCode
# we sum the Quantity in order to reach the quantities that were eventually bought from customers
df["CIDxSC Summed Quantity"] = df.groupby(["Customer ID", "StockCode"])["Quantity"].transform("sum")
df.shape

df.head()

# Keep only summed quantities with positive value
df = df[df["CIDxSC Summed Quantity"] > 0]

# Shape of df after removal of negative values
rows_after_negq_drop = df.shape[0]
df.shape

# Confirm that all negative summed quantities are removed
len(df[df["CIDxSC Summed Quantity"] <= 0])

# A number of rows with negative quantities remain after the removal of negative summed quantities. 
# These don't matter because they were already taken into account during the summation. 
# The ones that matter are those of column 'CIDxSC Summed Quantity', which have been dropped already.
df.shape
len(df[df["Quantity"] <= 0])

# Now drop the 'CIDxSC Summed Quantity' column: its only utility was for dropping the negative summed quantities
df.drop(["CIDxSC Summed Quantity"], axis=1, inplace=True)
df.head()

# Assert quantity is not null per customer for all customers
df["Quantity"].isnull().any()

# Assert price > 0 for all prices and stockcodes
len(df[df["Price"] < 0])

# Remove time from InvoiceDate column
df["InvoiceDate"] = pd.to_datetime(df["InvoiceDate"]).dt.date
df.head()

# Review how many cancelled invoices remain in the dataset after the above modifications
len(df[df["Invoice"].str.contains("C")])

# Check the cancelled invoices: 
# we will not remove them although their quantities are negative
# because they have been already been filtered through the `CIDxSC Summed Quantity` analysis
# where all negative summed quantities were already dropped
df[df["Invoice"].str.contains("C")].head()

# Export df to excel: this is the final form of the dataset that will be used for RFM analysis
df_clean = df.to_excel(r'C:\Users\argyr\OneDrive\Υπολογιστής\Diplvm\Giftware dataset\clean_data.xlsx', index=False)
