==> Notebooks/05.compare_all_three_methods_by_grouping_into_four.ipynb <==
# Import libraries
import numpy as np 
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import datetime as dt

import feature_engine
from feature_engine.outliers import Winsorizer

import warnings
warnings.filterwarnings("ignore")

import plotly.express as px

%matplotlib ipympl


rfm_perc = pd.read_csv(r'C:\Users\argyr\OneDrive\Υπολογιστής\Diplvm\Giftware dataset\rfm_sc_perc.csv')
rfm_km   = pd.read_csv(r'C:\Users\argyr\OneDrive\Υπολογιστής\Diplvm\Giftware dataset\rfm_sc_km.csv')
rfm_km4  = pd.read_csv(r'C:\Users\argyr\OneDrive\Υπολογιστής\Diplvm\Giftware dataset\df_rfm_k4.csv')

rfm_perc.head()

# This is our grouping of 11 segments into 4 groups
segments_group_0 = ["champions", "loyal"]
segments_group_1 = ["potential_loyalist", "new_customers", "promising", "need_attention"]
segments_group_2 = ["about_to_sleep", "cannot_lose_them_but_losing", "at_risk"]
segments_group_3 = ["hibernating", "lost"]

# Compute the 4 groups for rfm_perc
rfm_perc_g0 = rfm_perc[rfm_perc['segment_name'].isin(segments_group_0)]
rfm_perc_g1 = rfm_perc[rfm_perc['segment_name'].isin(segments_group_1)]
rfm_perc_g2 = rfm_perc[rfm_perc['segment_name'].isin(segments_group_2)]
rfm_perc_g3 = rfm_perc[rfm_perc['segment_name'].isin(segments_group_3)]

# Compute the 4 groups for rfm_km
rfm_km_g0 = rfm_km[rfm_km['segment_name'].isin(segments_group_0)]
rfm_km_g1 = rfm_km[rfm_km['segment_name'].isin(segments_group_1)]
rfm_km_g2 = rfm_km[rfm_km['segment_name'].isin(segments_group_2)]
rfm_km_g3 = rfm_km[rfm_km['segment_name'].isin(segments_group_3)]

rfm_perc_g0.shape[0],rfm_perc_g1.shape[0],rfm_perc_g2.shape[0],rfm_perc_g3.shape[0]

rfm_km_g0.shape[0],rfm_km_g1.shape[0],rfm_km_g2.shape[0],rfm_km_g3.shape[0]

def rfm_values(df):

    df_new = df.groupby(['Cluster']).agg({
        'Recency': 'mean',
        'Frequency': 'mean',
        'Monetary': ['mean', 'count']
    }).round(0)
    
    return df_new

print(rfm_values(rfm_km4).loc[0, ('Monetary', 'count')], rfm_values(rfm_km4).loc[1, ('Monetary', 'count')], rfm_values(rfm_km4).loc[2, ('Monetary', 'count')], rfm_values(rfm_km4).loc[3, ('Monetary', 'count')])

def compute_total_value(sizes, weights):
    return sum([a * b for a, b in zip(sizes, weights)])

weights = [1,2,3,4]
rfm_perc_gsz = [rfm_perc_g0.shape[0],rfm_perc_g1.shape[0],rfm_perc_g2.shape[0],rfm_perc_g3.shape[0]]
rfm_km_gsz = [rfm_km_g0.shape[0],rfm_km_g1.shape[0],rfm_km_g2.shape[0],rfm_km_g3.shape[0]]
rfm_km4_gsz = [rfm_values(rfm_km4).loc[0, ('Monetary', 'count')], rfm_values(rfm_km4).loc[1, ('Monetary', 'count')], rfm_values(rfm_km4).loc[2, ('Monetary', 'count')], rfm_values(rfm_km4).loc[3, ('Monetary', 'count')]]

print("Percentiles score:   ", compute_total_value(rfm_perc_gsz, weights))
print("Column k-means score:", compute_total_value(rfm_km_gsz, weights))
print("RFM k-means score:   ", compute_total_value(rfm_km4_gsz, weights))
